+++ CMakeLists.txt
@@ -29,9 +29,9 @@ target_include_directories (sandbox_executable PRIVATE ${LLVM_INCLUDE_DIRS})
 # add compile definitions/ macros to sandbox
 target_compile_definitions (sandbox_executable PRIVATE ${LLVM_DEFINITIONS_LIST})
 
-target_include_directories (sandbox_executable PRIVATE sandbox)
+target_include_directories (sandbox_executable PRIVATE sandbox kaleidoscope)
 
-add_subdirectory (sandbox)
+# add_subdirectory (sandbox)
 
 # kaleidoscope implementation
-add_subdirectory(kaleidoscope)
+add_subdirectory (kaleidoscope)
