+++ kaleidoscope/kaleidoscope_example.txt
@@ -1,0 +1,7 @@
+def fib(x)
+    if x < 3 then
+      1
+    else
+      fib(x - 1) + fib(x - 2)
+
+fib(40)
