+++ kaleidoscope/lexer/CMakeLists.txt
@@ -1,0 +1,7 @@
+target_sources (sandbox_executable PRIVATE
+                Reader.cpp
+                Reader.hpp
+                Lexer.cpp
+                Lexer.hpp
+                Token.hpp
+)
