+++ kaleidoscope/lexer/Lexer.hpp
@@ -1,0 +1,18 @@
+#pragma once
+
+#include <string>
+#include "lexer/Reader.hpp"
+#include "lexer/Token.hpp"
+
+class Lexer {
+public:
+    Lexer(Reader &r);
+    ~Lexer() = default;
+    Token getToken();
+
+private:
+    Reader& reader;
+    int lastCharacter;
+    std::string identifierStr;
+    double valNum;
+};
