+++ sandbox/CMakeLists.txt
@@ -1,3 +1,1 @@
 ﻿target_sources (sandbox_executable PRIVATE hello_llvm.cpp)
-
-add_subdirectory (lexer)
